pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-pipeline"
version = "0.1.0"
description = "Simple RAG pipeline using LangChain and local LLM"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-text-splitters>=0.3.0",
    "langchain-openai>=0.3.0",
    "openai>=1.3.0",
    "pypdf>=3.17.1",
    "python-dotenv==1.0.0",
    "faiss-cpu>=1.13.2",
    "numpy>=2.4.3"
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
